+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Fixed
 ^^^^^
 
 * Include support for deleters when ignoring property decorators (`#126 <https://github.com/econchick/interrogate/issues/126>_`).
+* Support floats for `--fail-under` values (`#114 <https://github.com/econchick/interrogate/issues/114>`_).
 
 Removed
 ^^^^^^^